pyproject.toml
[tool.poetry]
name = "edge_sim_py"
version = "1.1.0"
description = "Python-based modeling and simulation framework for Edge Computing resource management policies"
authors = ["Paulo Severo <paulo.severo@edu.pucrs.br>"]

[tool.poetry.dependencies]
python = ">=3.10,<3.11"
Mesa = "^1.0.0"
networkx = "2.6.2"
msgpack = "^1.0.4"
dm-acme = "^0.4.0"
protobuf = "3.20.3"
dm-launchpad = "0.5.0"
dm-reverb = {version = "0.8.0", extras = ["tensorflow"]}
jax = "^0.4.31"
jaxlib = "^0.4.31"
numpy = "<2"
tensorflow = "2.9.0"
dm-sonnet = "^2.0.2"
tensorflow-probability = "0.17"
gym = "^0.26.2"
trfl = "^1.2.0"

[tool.poetry.dev-dependencies]
black = "^22.8.0"
mkdocs = "^1.3.1"
mkdocs-material = "^8.4.2"
mkdocstrings = "^0.19.0"
mkdocstrings-python = "^0.7.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 130